Add destroy handling of Target

+++ server/apps/targets/logic/target_change_balance.py
@@ -1,10 +1,11 @@
 from datetime import datetime
+from decimal import Decimal
 
 from apps.targets.models.target import Target
 from apps.targets.models.target_change_balance import TargetChangeBalance
 
 
-def create_change_balance_now(target: Target, amount: int) -> TargetChangeBalance:
+def create_change_balance_now(target: Target, amount: Decimal) -> TargetChangeBalance:
     change_balance = TargetChangeBalance()
     change_balance.target = target
     change_balance.amount = amount
